Use triggers when calling update-initramfs in postinst and postrm
One complication with the interaction between live-initramfs and
update-initramfs is completeness - in a normal Debian system, we do not
mind that not all the initrds are updated, so one would simply call
"update-initramfs -u" when upgrading (for example) mdadm.

However, live-initramfs needs to update *all* initrds (ie. '-u -k all'),
but calling this would not use the trigger functionality as
'update-initramfs' will only accept triggering with a single '-u' argument.
This is sensible behaviour, as the trigger callback cannot (and shouldn't)
be parameterised. The result is this is that we must detect how many kernel
flavours we have and revert to the non-triggered call ('-u -k all') if we
have more than one.

My tests show that this is behaviour is sound, even in the slightly
contrived example of updating live-initramfs whilst installing a second
kernel, regardless of the interleaving of the calls to postinst.

We additionally attempt the same logic in the postrm so that upgrading
live-initramfs (for example, in chroot_local-packages) does not result in
two calls expensive calls in update-initramfs.

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>