Bugfix 5685: Allow live script to mount root from a CCISS RAID controller.
In searching for a device to mount the root filesystem from, the
script calls the is_nice_device() function to determine whether each
device should be tried.  It uses the common naming conventions to
identify disk devices, but did not recognize the naming convention
used by the CCISS RAID controller.  This change adds CCISS device
names to the list of valid names.