Set up networking before doing netbooting.
This fixes using fetch= to load filesystem.squashfs over HTTP.

You can tell do_netmount in scripts/live needs DEVICE to be set,
but the only place it is ever set is in configure_networking()
and that's not called anywhere in the code. That's a bug.

This is the essentially the same fix used in recent versions of this
package: at the top of do_netmount there is a call to do_netsetup which
is pretty much a renamed configure_networking.

I can't help but wonder if this ever worked at all in Debian. Anyway.